pyproject.toml
[tool.poetry]
name = "nsfw_detector"
version = "0.1.0"
description = ""
authors = ["Jay Dhanwant <jay@gobazzinga.io>"]

[tool.poetry.dependencies]
python = "^3.8"
cachetools = "5.3.3"
certifi = "2024.6.2"
charset-normalizer = "3.3.2"
google-api-core = "2.19.1"
google-auth = "2.31.0"
google-cloud-bigquery = "3.25.0"
google-cloud-core = "2.4.1"
google-crc32c = "1.5.0"
google-resumable-media = "2.7.1"
googleapis-common-protos = "1.63.2"
grpcio = "1.65.2"
grpcio-tools = "1.65.2"
grpcio-status = "1.65.2"
idna = "3.7"
packaging = "24.1"
proto-plus = "1.24.0"
protobuf = "5.27.2"
pyasn1 = "0.6.0"
pyasn1-modules = "0.4.0"
python-dateutil = "2.9.0.post0"
requests = "2.32.3"
rsa = "4.9"
six = "1.16.0"
urllib3 = "2.2.2"
mlxtend = "0.23.1"
upstash-vector = "0.5.0"
pyyaml = "*"
grpcio-reflection = "1.65.2"
pandas = "2.2.2"
PyJWT = "2.9.0"
torch = "2.4.0"
torchvision = "*"
opencv-python = "4.10.0.84"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
